+++ src/constant/columns/transactionColumns.js
@@ -1,10 +1,11 @@
-import { formatDate } from "../helper";
+import { formatAmount, formatDate, formatTransactionType } from "../helper";
 
 const transactionColumns = [
-  { title: "Image", key: "url" },
-  { title: "Title", key: "title" },
-  { title: "Description", key: "description" },
+  { title: "Name", key: "name" },
+  { title: "Category", key: "category" },
   { title: "Date", key: "date", function: formatDate },
+  { title: "Amount", key: "amount", function: formatAmount },
+  { title: "Type", key: "type", function: formatTransactionType },
   { title: "Actions", key: "action" },
 ];
