+++ src/constant/inputs/transactionInputs.js
@@ -5,13 +5,7 @@ const transactionInputs = [
     type: "text",
     placeholder: "Bills.",
     label: "name",
-  },
-  {
-    id: 3,
-    name: "description",
-    type: "text",
-    placeholder: "description...",
-    label: "description",
+    className: "order-first",
   },
   {
     id: 2,
@@ -19,6 +13,18 @@ const transactionInputs = [
     type: "date",
     label: "date",
     max: new Date().toISOString().split("T")[0],
+    className: "order-3",
+  },
+  {
+    id: 3,
+    name: "amount",
+    type: "number",
+    placeholder: "200...",
+    label: "amount",
+    min: 1,
+    pattern: "/^[1-9]d*$/",
+    errorMessage: "Amount must be greater than 01",
+    className: "order-last",
   },
 ];
