+++ src/hooks/useTransaction.js
@@ -1,8 +1,20 @@
 import React, { useState } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { toast } from "react-toastify";
+import {
+  addTransactionAPI,
+  deleteTransactionAPI,
+  getTransactionAPI,
+  updateTransactionAPI,
+} from "../config/service/firebase/transaction";
+import {
+  addTransactionReducer,
+  deleteTransactionReducer,
+  getTransactionReducer,
+  updateTransactionReducer,
+} from "../feature/transaction/transactionSlice";
 
-const useTransition = () => {
+const useTransaction = () => {
   const [initLoading, setInitLoading] = useState(false);
   const [loading, setLoading] = useState(false);
 
@@ -13,15 +25,15 @@ const useTransition = () => {
     try {
       setInitLoading(true);
       let transaction = [];
-      // let response = await getExpenseAPI(userId);
-      // response.forEach((element) => {
-      //   transaction.push({
-      //     docId: element.id,
-      //     ...element.data(),
-      //     amount: +element.data()?.amount,
-      //   });
-      // });
-      // dispatch(getExpenseReducer(transaction));
+      let response = await getTransactionAPI(userId);
+      response.forEach((element) => {
+        transaction.push({
+          docId: element.id,
+          ...element.data(),
+          amount: +element.data()?.amount,
+        });
+      });
+      dispatch(getTransactionReducer(transaction));
     } catch (error) {
       toast.error(error?.message);
     } finally {
@@ -32,8 +44,8 @@ const useTransition = () => {
   const useAddTransaction = async (body, setValues) => {
     try {
       setLoading(true);
-      // let response = await addExpenseAPI(body);
-      // dispatch(addExpenseReducer({ ...body, docId: response?.id }));
+      let response = await addTransactionAPI(body);
+      dispatch(addTransactionReducer({ ...body, docId: response?.id }));
       setValues({});
       toast.success("Create successfully!");
     } catch (error) {
@@ -46,8 +58,8 @@ const useTransition = () => {
   const useUpdateTransaction = async (body, docId) => {
     try {
       setLoading(true);
-      // await updateExpenseAPI(body, docId);
-      // dispatch(updateExpenseReducer({ ...body, docId: docId }));
+      await updateTransactionAPI(body, docId);
+      dispatch(updateTransactionReducer({ ...body, docId: docId }));
       toast.success("Update successfully!");
     } catch (error) {
       toast.error(error?.message);
@@ -58,8 +70,8 @@ const useTransition = () => {
   const useDeleteTransaction = async (docId) => {
     try {
       setLoading(true);
-      // await deleteExpenseAPI(docId);
-      // dispatch(deleteExpenseReducer(docId));
+      await deleteTransactionAPI(docId);
+      dispatch(deleteTransactionReducer(docId));
       toast.success("Delete successfully!");
     } catch (error) {
       toast.error(error?.message);
@@ -78,4 +90,4 @@ const useTransition = () => {
   };
 };
 
-export default useTransition;
+export default useTransaction;
