+++ src/screens/sources-transaction/style.css
@@ -1,0 +1,33 @@
+.file-upload-wrapper {
+    position: relative;
+    width: 100%;
+    margin: 0 auto;
+}
+
+.file-input {
+    width: 100%;
+    height: 100%;
+    opacity: 0;
+    position: absolute;
+    left: 0;
+    top: 0;
+    cursor: pointer;
+}
+
+.file-label {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    padding: 15px;
+    border: 2px dashed #696cff ;
+    border-radius: 5px;
+    font-size: 16px;
+    color: #696cff ;
+    cursor: pointer;
+    height: 186px;
+}
+
+.file-label i {
+    margin-right: 8px;
+    font-size: 20px;
+}
